Implement add and update functionality for orders
Added POST and PUT HTTP methods to the OrderController for creating and updating orders, respectively. A significant portion of the new code is in OrderService, where orders are processed. This also accounts for transactions within the order, using the data transfer object (DTO) pattern to transfer data between processes. Data.sql file was altered to reflect these changes. Adjusted OrderDTO to include an OrderLineDTO and status.